backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.1.0
whitenoise==6.6.0
redis==5.0.0
django-redis==5.4.0
torch>=2.0.0
sentence-transformers>=2.2.0
transformers>=4.21.0
